+++ addons/antennas/CfgVehicles.hpp
@@ -2,6 +2,6 @@ class CfgVehicles {
     class Land_Communication_F;
 
     class TFAR_Land_Communication_F : Land_Communication_F {
-        displayName = "Communication Tower (TFAR)";
+        displayName = CSTRING(TowerName);
     };
 };
