+++ addons/antennas/XEH_PREP.sqf
@@ -1,0 +1,4 @@
+PREP(deleteRadioTower);
+PREP(initRadioTower);
+PREP(pluginAddRadioTower);
+PREP(pluginRemoveRadioTower);
