--- addons/antennas/XEH_postInit.sqf
@@ -1,3 +1,0 @@
-#include "script_component.hpp"
-
-if (!isServer) exitWith {};
