+++ addons/antennas/functions/fnc_initRadioTower.sqf
@@ -17,10 +17,12 @@
         NOTHING
 
     Example:
-        [_this,50000] call TFAR_fnc_initRadioTower;
+        [_this,50000] call TFAR_antennas_fnc_initRadioTower;
 */
 params ["_tower","_range"];
-diag_log ["TFAR_antennas_fnc_initRadioTower",_this];
+
+TRACE_1("TFAR_antennas_fnc_initRadioTower",_this);
+
 [GVAR(radioTowerList), _tower, _range] call CBA_fnc_hashSet;
 
-[DFUNC(pluginAddRadioTower),[_tower]] call CBA_fnc_execNextFrame; //netID may still be 0:0 directly at Init EH
+[FUNC(pluginAddRadioTower),[_tower]] call CBA_fnc_execNextFrame; //netID may still be 0:0 directly at Init EH
