+++ addons/antennas/functions/fnc_pluginAddRadioTower.sqf
@@ -18,11 +18,11 @@
         NOTHING
 
     Example:
-        [_tower1,_tower2] call TFAR_fnc_pluginAddRadioTower;
+        [_tower1,_tower2] call TFAR_antennas_fnc_pluginAddRadioTower;
 */
 
-//params get's weird with array Input so I just don't use it.
-diag_log ["TFAR_antennas_fnc_pluginAddRadioTower",_this];
+TRACE_1("TFAR_antennas_fnc_pluginAddRadioTower", _this);
+
 for "_y" from 0 to (count _this)-1 step 50 do { //Only 50 per call to not exceed max message length
     private _towersToProcess = (_this select [_y,50]);
     private _towerData = _towersToProcess apply {
@@ -30,7 +30,9 @@ for "_y" from 0 to (count _this)-1 step 50 do { //Only 50 per call to not exceed
         private _netID = netID _x;
         ([_netID,_range,position _x] joinString ";")
     };
-    diag_log _towerData;
-    systemChat "sendBatch";
+    #ifdef DEBUG_MODE_FULL
+        diag_log _towerData;
+        systemChat "sendBatch";
+    #endif
     "task_force_radio_pipe" callExtension format["RadioTwrAdd	%1~", _towerData joinString TF_new_line];
 };
